Refactor CollisionEvent class into separate file
